pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sec-filing-intelligence-copilot"
version = "0.1.0"
description = "Production-style source-grounded RAG system over SEC filings."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Shreyash Kondakindi" }
]
dependencies = [
  "beautifulsoup4>=4.12",
  "chromadb>=1.5.5,<2",
  "numpy>=1.26.4,<2",
  "openai>=2.29.0,<3",
  "pydantic>=2.12.5,<3",
  "PyYAML>=6.0.3,<7",
  "python-dotenv>=1.2.2,<2",
  "requests>=2.33.0,<3",
  "sentence-transformers>=5.3.0,<6",
  "tiktoken>=0.12.0,<1",
  "torch>=2.2.2,<3",
  "transformers>=4.57.6,<6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "live_sec: hits live SEC endpoints and is skipped by default",
]
